Rename to role_assignment_agent
